pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "narativ-network"
version = "0.1.0"
description = "Self-running 24/7 narrative TV channel: ingest, process, schedule, playout."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "pydantic>=2.6",
    "tomli>=2.0; python_version<'3.11'",
    "google-api-python-client>=2.120",
    "google-auth>=2.28",
    "google-auth-httplib2>=0.2",
    "apscheduler>=3.10",
    "rich>=13.7",
    "typer>=0.9",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
nn = "narativ_network.cli:app"

[tool.setuptools.packages.find]
include = ["narativ_network*"]

[tool.ruff]
line-length = 100
target-version = "py311"
